LayoutTests/http/tests/security/cross-frame-access-call.html

<html>
<head>
    <script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe src="http://localhost:8000/security/resources/cross-frame-iframe-for-get-test.html" style=""></iframe>
<pre id="console"></pre>
<script>

window.targetWindow = frames[0];

window.onload = function()
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    log("\n----- tests for calling methods of another frame using Function.call -----\n");

    // undefined value indicates failure
    shouldBe("setTimeout.call(targetWindow, 'void(0);', 0)", undefined);
    shouldBe("setInterval.call(targetWindow, 'void(0);', 0)", undefined);

    // these always return undefined so we use the error console to detect failure
    addEventListener.call(targetWindow, "load", null, false);
    removeEventListener.call(targetWindow, "load", null, false);

    // Work around DRT bug that causes subsequent tests to fail.
    window.stop();
}
</script>
</body>
</html>
